requirements_gpu.txt
tensorflow-gpu
pandas
xlrd
openpyxl
scikit-allel
myvariant